changed ebpftelemetry metrics from gauge to counter
