+++ pkg/ebpf/telemetry/errors_collector_linux.go
@@ -9,10 +9,20 @@ package telemetry
 
 import (
 	"fmt"
+	"golang.org/x/sys/unix"
+	"syscall"
 
 	"github.com/prometheus/client_golang/prometheus"
 
 	"github.com/DataDog/datadog-agent/pkg/util/log"
+)
+
+const (
+	maxErrno    = 64
+	maxErrnoStr = "other"
+
+	ebpfMapTelemetryNS    = "ebpf_maps"
+	ebpfHelperTelemetryNS = "ebpf_helpers"
 )
 
 // EBPFErrorsCollector implements the prometheus Collector interface
@@ -21,6 +31,8 @@ type EBPFErrorsCollector struct {
 	*EBPFTelemetry
 	ebpfMapOpsErrorsGauge *prometheus.Desc
 	ebpfHelperErrorsGauge *prometheus.Desc
+	//we can use one map for both map errors and ebpf helpers errors, as the keys are different
+	lastValues map[string]uint64 // used to calculate the delta of the error counters
 }
 
 // NewEBPFErrorsCollector initializes a new Collector object for ebpf helper and map operations errors
@@ -35,6 +47,7 @@ func NewEBPFErrorsCollector() prometheus.Collector {
 		},
 		ebpfMapOpsErrorsGauge: prometheus.NewDesc(fmt.Sprintf("%s__errors", ebpfMapTelemetryNS), "Failures of map operations for a specific ebpf map reported per error.", []string{"map_name", "error"}, nil),
 		ebpfHelperErrorsGauge: prometheus.NewDesc(fmt.Sprintf("%s__errors", ebpfHelperTelemetryNS), "Failures of bpf helper operations reported per helper per error for each probe.", []string{"helper", "probe_name", "error"}, nil),
+		lastValues:            make(map[string]uint64),
 	}
 }
 
@@ -57,11 +70,15 @@ func (e *EBPFErrorsCollector) Collect(ch chan<- prometheus.Metric) {
 				log.Debugf("failed to get telemetry for probe:key %s:%d\n", probeName, k)
 				continue
 			}
-			for indx, helperName := range helperNames {
-				base := maxErrno * indx
+			for index, helperName := range helperNames {
+				base := maxErrno * index
 				if count := getErrCount(hval.Count[base : base+maxErrno]); len(count) > 0 {
 					for errStr, errCount := range count {
-						ch <- prometheus.MustNewConstMetric(e.ebpfHelperErrorsGauge, prometheus.GaugeValue, float64(errCount), helperName, probeName, errStr)
+						errorsDelta := float64(errCount - e.lastValues[errStr])
+						if errorsDelta > 0 {
+							ch <- prometheus.MustNewConstMetric(e.ebpfHelperErrorsGauge, prometheus.CounterValue, errorsDelta, helperName, probeName, errStr)
+						}
+						e.lastValues[errStr] = errCount
 					}
 				}
 			}
@@ -78,9 +95,31 @@ func (e *EBPFErrorsCollector) Collect(ch chan<- prometheus.Metric) {
 			}
 			if count := getErrCount(val.Count[:]); len(count) > 0 {
 				for errStr, errCount := range count {
-					ch <- prometheus.MustNewConstMetric(e.ebpfMapOpsErrorsGauge, prometheus.GaugeValue, float64(errCount), m, errStr)
+					errorsDelta := float64(errCount - e.lastValues[errStr])
+					if errorsDelta > 0 {
+						ch <- prometheus.MustNewConstMetric(e.ebpfMapOpsErrorsGauge, prometheus.CounterValue, errorsDelta, m, errStr)
+					}
+					e.lastValues[errStr] = errCount
 				}
 			}
 		}
 	}
+}
+
+func getErrCount(v []uint64) map[string]uint64 {
+	errCount := make(map[string]uint64)
+	for i, count := range v {
+		if count == 0 {
+			continue
+		}
+
+		if (i + 1) == maxErrno {
+			errCount[maxErrnoStr] = count
+		} else if name := unix.ErrnoName(syscall.Errno(i)); name != "" {
+			errCount[name] = count
+		} else {
+			errCount[syscall.Errno(i).Error()] = count
+		}
+	}
+	return errCount
 }
