+++ pkg/ebpf/telemetry/errors_telemetry.go
@@ -10,28 +10,18 @@ package telemetry
 import (
 	"errors"
 	"fmt"
+	"golang.org/x/exp/slices"
 	"hash"
 	"hash/fnv"
 	"sync"
-	"syscall"
 
 	manager "github.com/DataDog/ebpf-manager"
 	"github.com/cilium/ebpf"
 	"github.com/cilium/ebpf/asm"
-	"golang.org/x/exp/slices"
-	"golang.org/x/sys/unix"
 
 	"github.com/DataDog/datadog-agent/pkg/ebpf/maps"
 	"github.com/DataDog/datadog-agent/pkg/network/ebpf/probes"
 	"github.com/DataDog/datadog-agent/pkg/util/kernel"
-)
-
-const (
-	maxErrno    = 64
-	maxErrnoStr = "other"
-
-	ebpfMapTelemetryNS    = "ebpf_maps"
-	ebpfHelperTelemetryNS = "ebpf_helpers"
 )
 
 const (
@@ -92,24 +82,6 @@ func (b *EBPFTelemetry) populateMapsWithKeys(m *manager.Manager) error {
 		return err
 	}
 	return nil
-}
-
-func getErrCount(v []uint64) map[string]uint64 {
-	errCount := make(map[string]uint64)
-	for i, count := range v {
-		if count == 0 {
-			continue
-		}
-
-		if (i + 1) == maxErrno {
-			errCount[maxErrnoStr] = count
-		} else if name := unix.ErrnoName(syscall.Errno(i)); name != "" {
-			errCount[name] = count
-		} else {
-			errCount[syscall.Errno(i).Error()] = count
-		}
-	}
-	return errCount
 }
 
 func buildMapErrTelemetryConstants(mgr *manager.Manager) []manager.ConstantEditor {
